+++ src/js/features/card-page-features/index.ts
@@ -1,4 +1,5 @@
 import TagMenu from "./tag-menu";
 import CardNicknameDisplay from "./card-nickname-display";
+import CardSearchLinks from "./card-search-links";
 
-export default [TagMenu, CardNicknameDisplay];
+export default [TagMenu, CardNicknameDisplay, CardSearchLinks];